+++ src/packages/BookmarksManager/src/components/BookmarksSidesheet/Modal/index.ts
@@ -1,4 +1,3 @@
-export * from './Modal';
 export * from './DeleteModalContent';
 export * from './EditModalContent';
 export * from './ShareModalContent';
